minor changes to db models and resources
