+++ api/models/student_course.py
@@ -4,13 +4,14 @@ class StudentCourse(db.Model):
     __tablename__= 'student_course'
     student_id = db.Column(db.Integer(), db.ForeignKey('students.student_id'), primary_key=True)
     course_id = db.Column(db.Integer(), db.ForeignKey('courses.course_id'), primary_key=True)
+    # students = db.relationship('Student', backref='student_course')
     #grade = db.Column(db.String(), nullable=False)
     #credit_unit = db.Column(db.Float(), nullable=False)
     #earned_credit = db.Column(db.Float(), nullable=True)
     #teacher = db.relationship('Teacher', secondary='courses', backref='student_courses')
     
     def __repr__(self):
-        return f"<Student Course ID {self.id}>"
+        return f"<Student Course ID {self.student_id}>"
     
     def save(self):
         db.session.add(self)
