+++ api/models/student_grades.py
@@ -1,0 +1,2 @@
+from ..db import db
+
